@@ -29,11 +29,14 @@
     "axios": "^0.16.1",
     "babel-preset-stage-1": "^6.1.18",
     "lodash": "^3.10.1",
-    "react": "^0.14.3",
-    "react-dom": "^0.14.3",
+    "prop-types": "^15.5.10",
+    "react": "^15.5.0",
+    "react-dom": "^15.3.2",
     "react-jsonschema-form": "^0.48.2",
     "react-redux": "4.3.0",
     "react-router": "^2.0.1",
-    "redux": "^3.0.4"
+    "redux": "^3.0.4",
+    "redux-form": "^6.7.0",
+    "redux-react-schema-form": "^1.1.3"
   }
 }
