@@ -1,14 +1,15 @@
 /* global $:true */
 
 import React, {
-  Component,
-  PropTypes
+  Component
 }                           from 'react';
+import PropTypes            from 'prop-types';
 import Form                 from "react-jsonschema-form";
 
 const schema = {
 	title: "Status",
   type: "object",
+  // required: ['status_state'],
   properties: {
     status_state: {
         type: "string",
@@ -26,6 +27,10 @@ const uiSchema = {
 	status_state: {
     "ui:widget": "radio",
   }
+}
+
+const formData = {
+  status_state: ''
 }
 
 const log = (type) => console.log.bind(console, type);
@@ -49,6 +54,14 @@ class WizardFormStatusPage extends React.Component {
   }
 
   render() {
+    
+    const validate = (formData, errors) => {
+      console.log("FormData: ", formData);
+      if (formData.status_state == '' || typeof formData.status_state == "undefined") {
+        errors.status_state.addError("The field is required");
+      }
+      return errors;
+    }
 
     return (
       <div>
@@ -56,6 +69,7 @@ class WizardFormStatusPage extends React.Component {
                 uiSchema={uiSchema}
                 onChange={log("changed")}
                 onSubmit={log("submitted")}
+                validate={validate}
                 onError={log("errors")} />
           <button type="button" className="previous" onClick={ this.props.previousPage }>
               Previous
@@ -68,8 +82,8 @@ class WizardFormStatusPage extends React.Component {
 }
 
 WizardFormStatusPage.propTypes = {
-  previousPage  : React.PropTypes.func,
-  nextPage      : React.PropTypes.func
+  previousPage  : PropTypes.func,
+  nextPage      : PropTypes.func
 }
 
 export default WizardFormStatusPage;
