@@ -1,9 +1,25 @@
 import {CATEGORY_SELECTED} from '../actions/index';
+import {FETCH_SAGA} from '../actions/index';
 
-export default function(state = null, action) {
+const initState = {
+	filter: '',
+	data: '',
+    currentPage: 1
+};
+
+export default function(state = initState, action) {
     switch (action.type) {
         case CATEGORY_SELECTED:
-            return action.payload;
+            return {
+                ...state,
+                data: action.payload
+            }
+
+        case FETCH_SAGA:
+        	return {
+                ...state,
+        		schemaData: action.payload
+        	}
     }
     return state;
 }
